+++ api/server.js
@@ -8,10 +8,12 @@ server.use(cors())
 const authRouter = require('../auth/auth-router')
 const usersRouter = require('../users/users-router.js')
 const postsRouter = require('../posts/posts-router')
+const homeRouter = require('../home/home-router')
 
 server.use('/api/auth', authRouter)
 server.use('/api/users', usersRouter)
 server.use('/api/posts', postsRouter)
+server.use('/api/home', homeRouter)
 
 server.get('/', (req, res) => {
     res.json('Lambda School Build Week - June 24-28 - Design Your Life')
