+++ home/home-model.js
@@ -1,0 +1,9 @@
+const db = require("../data/dbConfig");
+
+module.exports = {
+  getUserPosts,
+};
+
+function getUserPosts(id) {
+    return db("posts").where({ user_id: id });
+  }
